torrent/torrent_tracker/requirements.txt
asgiref==3.2.3
awscli==1.16.253
certifi==2019.11.28
colorama==0.4.1
Django==3.0.1
djangorestframework==3.11.0
docutils==0.15.2
jmespath==0.9.4
psycopg2==2.8.4
pyasn1==0.4.7
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
rsa==3.4.2
s3transfer==0.2.1
sqlparse==0.3.0
